Use regular division inside Scale Estimation (#3210)
### Changes

Compute division inside SE algorithm always as `a/b` instead of
`a*(1/b)` in some cases.

### Reason for changes

During implementation #2727 some choices were made regarding how
division operation is computed in order for the changes to be completely
aligned with the previous implementation. Namely, before #2727 some
divisions were computed as `a*(1/b)`, and this is currently still the
case.

The way these divisions are computed originally was not intended. Now,
all divisions are aligned to the `a/b` form.

Compression time and memory are roughly the same.

| Model | Compression | Compression Time Develop (sec.) | Compression
Time Branch (sec.) | Peak Memory Develop (MiB) | Peak Memory Branch
(MiB) |

|--------------|--------------|---------------------------------|--------------------------------|---------------------------|--------------------------|
| tiny-llama | int4, SE | 222* | 228* | 3030 | 3032 |
| phi4-mini | in4, SE | 789* | 790* | 10817 | 10768 |
| llama-3.1-8b | int4, SE | 1776* | 1801* | 17756 | 18224 |

*time column includes PT -> OV conversion time.

### Related tickets

163286

### Tests

- https://github.com/openvinotoolkit/nncf/actions/runs/13368886294
- NNCF/job/manual/job/post_training_weight_compression/324/
- OVVP validation :white_check_mark: