[gha] set timeout (#3396)
### Changes

Set a timeout for all jobs to prevent slow or stacked jobs from running
indefinitely.
Add a condition for the update_labels job to avoid failure when the API
changes check job is canceled or fails.

### Reason for changes

1. To catch slow or stacked jobs
2. To prevent the `update_labels` job from failing with `Error: no
artifacts found` when the `API changes check` job is canceled (e.g., due
to a new commit push or PR closure).