[ONNX] Add eliminate_nop_cast pass (#3376)
### Changes

Apply the `eliminate_nop_cast` optimization pass before building the
NNCF graph. This pass aims to eliminate no-op cast nodes in the graph. A
no-op cast is a cast that doesn't change the value, such as casting a
tensor to its own type.

### Reason for changes

Statistics cannot be collected after the no-op Cast nodes because such
nodes are removed from the ONNX inference graph during the session.

### Related tickets

Ref: 164211