[Weight Compression] Rework WC Algorithm to Return WC Params (#3636)
### Changes

Re-worked the `apply` method in WC algorithm to use an extra method to
return weights compression params such that the apply method is more
concise and only contains algorithm, quantization logic etc.

### Reason for changes

This is done so that OpenVINO quantizer can obtain the final collection
of weights compression parameters for all the nodes so that are to be
compressed.

---------

Co-authored-by: Daniil Lyakhov <daniil.lyakhov@intel.com>