[ONNX] Add MatMul for FBC/BC (#3657)
### Changes

Add support for the `MatMul` operation in the Fast Bias/Bias Correction
algorithm.

### Reason for changes

The Fast Bias/Bias Correction algorithm is not applied to the
`MatMul`->`Add` subgraph in which one of the inputs to the Add operation
is a constant. In this case, the `MatMul` operation is not considered as
having a bias, so the algorithm is not applied.

### Related tickets

Ref: 135433

### Tests

- test_update_bias_in_matmul_add
- text_examples:
[build](https://github.com/openvinotoolkit/nncf/actions/runs/18529108467)