[OpenVINO] Introduce a way to control resetting of the state of stateful OV models (#3714)
### Changes

Added `nncf.definitions.NNCF_DATASET_RESET_STATE_KEY` constant to
specify when to reset model state. This constant is used by OpenVINO
backend to control resetting of internal model state between model
inferences. This key can be added to a dataset sample input dictionary
with either `True` or `False` value. With `True` value, the model state
will be reset before inference on the corresponding sample, and with
`False` the state will not be reset.

For an example of usage please see
https://github.com/huggingface/optimum-intel/pull/1505.

### Reason for changes

Without this logic static quantization quality of stateful Whisper
models is poor because a state of a stateful model must be cleared with
the same schedule as it is done during calibration input data
collection.

### Related tickets

172705

### Tests

Added
`tests/openvino/native/test_engine.py::test_stateful_model_inference_with_controlled_resetting`.